# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "api autentia manager",
        "title": "Autentia Manager Documentation",
        "contact": {},
        "version": "1"
    },
    "paths": {
...
        "/logs/trx": {
            "post": {
                "security": [
                    {
                        "barerToken": []
                    }
                ],
                "description": "post logs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "log"
                ],
                "summary": "post Logs",
                "parameters": [
                    {
                        "description": "log",
                        "name": "log",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.logParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/routes.MessageResponse"
                        }
                    }
                }
            }
        },
        "/logs/{date}": {
            "get": {
                "security": [
                    {
                        "barerToken": []
                    }
                ],
                "description": "get logs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "log"
                ],
                "summary": "get Logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "date",
                        "name": "date",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "starthour",
                        "name": "starthour",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "finishhour",
                        "name": "finishhour",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.Log"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/routes.MessageResponse"
                        }
                    }
                }
            }
        },
        "/modules": {
            "get": {
                "security": [
                    {
                        "barerToken": []
                    }
                ],
                "description": "modules",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "module"
                ],
                "summary": "modules",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.Module"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/routes.MessageResponse"
                        }
                    }
                }
            }
        },
        "/modules/{name}": {
            "delete": {
                "security": [
                    {
                        "barerToken": []
                    }
                ],
                "description": "delete modules",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "module"
                ],
                "summary": "delete modules",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/routes.MessageResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "data.Function": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "string"
                },
                "module": {
                    "$ref": "#/definitions/data.Module"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "data.Log": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "module": {
                    "type": "string"
                },
                "params": {
                    "type": "string"
                },
                "person_dni": {
                    "type": "string"
                },
                "sensor": {
                    "type": "string"
                },
                "user_nickname": {
                    "type": "string"
                }
            }
        },
        "data.Module": {
            "type": "object",
            "properties": {
                "functions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/data.Function"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "routes.MessageResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "string"
                }
            }
        },
...
        "routes.logParams": {
            "type": "object",
            "required": [
                "input",
                "output",
                "resource"
            ],
            "properties": {
                "input": {
                    "type": "string"
                },
                "output": {
                    "type": "string"
                },
                "resource": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "barerToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}